create python test strategy
